refactor(menu): move vertical dots story to recipes
